+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * pool: changes in `RelayPool::remove_relay` behavior ([Yuki Kishimoto])
 * pool: deprecate `RelayPool::remove_all_relays` ([Yuki Kishimoto])
 * signer: bootstrap NIP-46 signer on demand ([Yuki Kishimoto])
+* bindings(nostr): adj. `tag` module ([Yuki Kishimoto])
 
 ### Added